# Optimizations based on commutativity of identical