--- a/cmd/modver/options.go
+++ b/cmd/modver/options.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/pkg/errors"
+	"github.com/bobg/errors"
 	"golang.org/x/mod/semver"
 )