--- a/cmd/modver/tags.go
+++ b/cmd/modver/tags.go
@@ -6,11 +6,11 @@ import (
 	"os"
 	"strings"
 
+	"github.com/bobg/errors"
 	"github.com/go-git/go-git/v5"
 	"github.com/go-git/go-git/v5/plumbing"
 	"github.com/go-git/go-git/v5/plumbing/object"
 	"github.com/go-git/go-git/v5/plumbing/storer"
-	"github.com/pkg/errors"
 	"golang.org/x/mod/semver"
 
 	"github.com/bobg/modver/v2"