--- a/modver_test.go
+++ b/modver_test.go
@@ -4,7 +4,6 @@ import (
 	"bufio"
 	"bytes"
 	"context"
-	"errors"
 	"fmt"
 	"io"
 	"log"
@@ -13,9 +12,19 @@ import (
 	"strings"
 	"testing"
 	"text/template"
+
+	"github.com/bobg/errors"
 )
 
 func TestCompare(t *testing.T) {
+	tbCompare(t)
+}
+
+func BenchmarkCompare(b *testing.B) {
+	tbCompare(b)
+}
+
+func tbCompare(tb testing.TB) {
 	cases := []struct {
 		dir  string
 		want ResultCode
@@ -30,15 +39,26 @@ func TestCompare(t *testing.T) {
 	}}
 
 	for _, c := range cases {
-		runtest(t, c.dir, c.want)
+		runtest(tb, c.dir, c.want)
 	}
 }
 
-func runtest(t *testing.T, typ string, want ResultCode) {
+func tbRun(tb testing.TB, name string, f func(testing.TB)) {
+	switch tb := tb.(type) {
+	case *testing.T:
+		tb.Run(name, func(t *testing.T) { f(tb) })
+	case *testing.B:
+		tb.Run(name, func(b *testing.B) { f(tb) })
+	}
+}
+
+func runtest(tb testing.TB, typ string, want ResultCode) {
+	b, _ := tb.(*testing.B)
+
 	tree := filepath.Join("testdata", typ)
 	entries, err := os.ReadDir(tree)
 	if err != nil {
-		t.Fatal(err)
+		tb.Fatal(err)
 	}
 	for _, entry := range entries {
 		if entry.IsDir() {
@@ -48,103 +68,117 @@ func runtest(t *testing.T, typ string, want ResultCode) {
 			continue
 		}
 		name := strings.TrimSuffix(entry.Name(), ".tmpl")
-		t.Run(fmt.Sprintf("%s/%s", typ, name), func(t *testing.T) {
-			tmpls, err := template.ParseFiles(filepath.Join(tree, entry.Name()))
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			tmpdir, err := os.MkdirTemp("", "modver")
-			if err != nil {
-				t.Fatal(err)
-			}
-			defer os.RemoveAll(tmpdir)
-
-			var (
-				olderTestDir = filepath.Join(tmpdir, "older")
-				newerTestDir = filepath.Join(tmpdir, "newer")
-			)
-
-			err = os.Mkdir(olderTestDir, 0755)
-			if err != nil {
-				t.Fatal(err)
-			}
-			err = os.Mkdir(newerTestDir, 0755)
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			var sawGomod bool
-			for _, tmpl := range tmpls.Templates() {
-				// Skip the top-level template
-				if strings.HasSuffix(tmpl.Name(), ".tmpl") {
-					continue
-				}
-
-				sawGomod = sawGomod || (filepath.Base(tmpl.Name()) == "go.mod")
-
-				parts := strings.Split(tmpl.Name(), "/")
-				if !strings.Contains(parts[len(parts)-1], ".") {
-					parts = append(parts, "x.go")
-				}
-
-				if len(parts) == 1 {
-					// Only a filename is given.
-					// Write it to both older and newer dirs.
-					buf := new(bytes.Buffer)
-					err = executeTmpl(tmpl, buf)
-					if err != nil {
-						t.Fatal(err)
-					}
-					for _, subdir := range []string{olderTestDir, newerTestDir} {
-						filename := filepath.Join(subdir, parts[0])
-						err = os.WriteFile(filename, buf.Bytes(), 0644)
+		tbRun(tb, fmt.Sprintf("%s/%s", typ, name), func(tb testing.TB) {
+			err := withTestDirs(tree, name, func(olderTestDir, newerTestDir string) {
+				if b != nil {
+					b.ResetTimer()
+					for i := 0; i < b.N; i++ {
+						_, err := CompareDirs(olderTestDir, newerTestDir)
 						if err != nil {
-							t.Fatal(err)
+							b.Fatal(err)
 						}
 					}
-					continue
+					return
 				}
 
-				if len(parts) > 1 {
-					dirparts := append([]string{tmpdir}, parts[:len(parts)-1]...)
-					dirname := filepath.Join(dirparts...)
-					err = os.MkdirAll(dirname, 0755)
-					if err != nil {
-						t.Fatal(err)
-					}
-				}
-				fileparts := append([]string{tmpdir}, parts...)
-				filename := filepath.Join(fileparts...)
-				err = executeTmplToFile(tmpl, filename)
+				got, err := CompareDirs(olderTestDir, newerTestDir)
 				if err != nil {
-					t.Fatal(err)
+					tb.Fatal(err)
 				}
-			}
-			if !sawGomod {
-				buf := new(bytes.Buffer)
-				fmt.Fprintf(buf, "module %s\n\ngo 1.18\n", name)
-				err = os.WriteFile(filepath.Join(olderTestDir, "go.mod"), buf.Bytes(), 0644)
-				if err != nil {
-					t.Fatal(err)
+				if got.Code() != want {
+					tb.Errorf("want %s, got %s", want, got)
+				} else {
+					tb.Log(got)
 				}
-				err = os.WriteFile(filepath.Join(newerTestDir, "go.mod"), buf.Bytes(), 0644)
-				if err != nil {
-					t.Fatal(err)
-				}
-			}
-
-			got, err := CompareDirs(olderTestDir, newerTestDir)
+			})
 			if err != nil {
-				t.Fatal(err)
-			}
-			if got.Code() != want {
-				t.Errorf("want %s, got %s", want, got)
-			} else {
-				t.Log(got)
+				tb.Fatal(err)
 			}
 		})
 	}
+}
+
+func withTestDirs(tree, name string, f func(olderTestDir, newerTestDir string)) error {
+	tmpls, err := template.ParseFiles(filepath.Join(tree, name+".tmpl"))
+	if err != nil {
+		return errors.Wrap(err, "parsing templates")
+	}
+
+	tmpdir, err := os.MkdirTemp("", "modver")
+	if err != nil {
+		return errors.Wrap(err, "creating temp dir")
+	}
+	defer os.RemoveAll(tmpdir)
+
+	var (
+		olderTestDir = filepath.Join(tmpdir, "older")
+		newerTestDir = filepath.Join(tmpdir, "newer")
+	)
+
+	if err = os.Mkdir(olderTestDir, 0755); err != nil {
+		return errors.Wrap(err, "creating older test dir")
+	}
+	if err = os.Mkdir(newerTestDir, 0755); err != nil {
+		return errors.Wrap(err, "creating newer test dir")
+	}
+
+	var sawGomod bool
+	for _, tmpl := range tmpls.Templates() {
+		// Skip the top-level template
+		if strings.HasSuffix(tmpl.Name(), ".tmpl") {
+			continue
+		}
+
+		sawGomod = sawGomod || (filepath.Base(tmpl.Name()) == "go.mod")
+
+		parts := strings.Split(tmpl.Name(), "/")
+		if !strings.Contains(parts[len(parts)-1], ".") {
+			parts = append(parts, "x.go")
+		}
+
+		if len(parts) == 1 {
+			// Only a filename is given.
+			// Write it to both older and newer dirs.
+			buf := new(bytes.Buffer)
+			if err = executeTmpl(tmpl, buf); err != nil {
+				return errors.Wrap(err, "executing template")
+			}
+			for _, subdir := range []string{olderTestDir, newerTestDir} {
+				filename := filepath.Join(subdir, parts[0])
+				if err = os.WriteFile(filename, buf.Bytes(), 0644); err != nil {
+					return errors.Wrapf(err, "writing file %s", filename)
+				}
+			}
+			continue
+		}
+
+		if len(parts) > 1 {
+			dirparts := append([]string{tmpdir}, parts[:len(parts)-1]...)
+			dirname := filepath.Join(dirparts...)
+			if err = os.MkdirAll(dirname, 0755); err != nil {
+				return errors.Wrapf(err, "creating dir %s", dirname)
+			}
+		}
+		fileparts := append([]string{tmpdir}, parts...)
+		filename := filepath.Join(fileparts...)
+		if err = executeTmplToFile(tmpl, filename); err != nil {
+			return errors.Wrapf(err, "executing template to file %s", filename)
+		}
+	}
+	if !sawGomod {
+		buf := new(bytes.Buffer)
+		fmt.Fprintf(buf, "module %s\n\ngo 1.18\n", name)
+		if err = os.WriteFile(filepath.Join(olderTestDir, "go.mod"), buf.Bytes(), 0644); err != nil {
+			return errors.Wrap(err, "writing older go.mod")
+		}
+		if err = os.WriteFile(filepath.Join(newerTestDir, "go.mod"), buf.Bytes(), 0644); err != nil {
+			return errors.Wrap(err, "writing newer go.mod")
+		}
+	}
+
+	f(olderTestDir, newerTestDir)
+
+	return nil
 }
 
 func executeTmpl(tmpl *template.Template, w io.Writer) error {